+++ internal/domain/category/repository/categories.go
@@ -9,6 +9,8 @@ import (
 	"personal-finance/internal/model"
 )
 
+const defaultIDCategory = "default_category_id"
+
 type Repository interface {
 	Add(ctx context.Context, category model.Category, userID string) (model.Category, error)
 	FindAll(ctx context.Context, userID string) ([]model.Category, error)
@@ -40,7 +42,10 @@ func (p PgRepository) Add(_ context.Context, category model.Category, userID str
 func (p PgRepository) FindAll(_ context.Context, userID string) ([]model.Category, error) {
 	var categories []model.Category
 	result := p.Gorm.Where("categories.user_id=?", userID).
-		Preload("SubCategories").
+		Or("categories.user_id=?", defaultIDCategory).
+		Preload("SubCategories",
+			p.Gorm.Where(`"sub_categories"."user_id"=?`, userID).
+				Or(`"sub_categories"."user_id"=?`, defaultIDCategory)).
 		Find(&categories)
 	if err := result.Error; err != nil {
 		return []model.Category{}, err